# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Khadamatak SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, rejected, cancelled]
        - name: method
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentsListResponse"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create payment (manual, starts as pending)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /payments/{id}/confirm:
    post:
      operationId: confirmPayment
      tags: [payments]
      summary: Confirm a pending payment (owner/manager/accountant only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /payments/{id}/reject:
    post:
      operationId: rejectPayment
      tags: [payments]
      summary: Reject a pending payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectPaymentBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
...
components:
  schemas:
...
    Payment:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: "YER"
        method:
          type: string
          enum: [cash, kuraimi, jawali, bank_transfer, other]
        status:
          type: string
          enum: [pending, confirmed, rejected, cancelled]
        contactName:
          type: string
        reference:
          type: string
        notes:
          type: string
        confirmedByName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, amount, method, status]
    PaymentResponse:
      type: object
      properties:
        payment:
          $ref: "#/components/schemas/Payment"
    PaymentsListResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        totalConfirmed:
          type: number
        totalPending:
          type: number
    CreatePaymentBody:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [cash, kuraimi, jawali, bank_transfer, other]
        contactId:
          type: string
        orderId:
          type: string
        reference:
          type: string
        notes:
          type: string
      required: [amount, method]
    RejectPaymentBody:
      type: object
      properties:
        reason:
          type: string
...